Debouncing work including IO loop rewrite

+++ Atm_element.cpp
@@ -135,7 +135,7 @@ bool Atm_element::idle( uint32_t maximum ) { // maximum time idle
 }
 
 
-Atm_element& Atm_element::debounce( uint8_t b, uint16_t r, uint8_t m ) {
+Atm_element& Atm_element::debounce( uint8_t b, uint16_t r, uint16_t m ) {
   playfield->debounce( switchno, b, r, m );
   return *this;
 }
